backend/requirements.txt
# Web 框架
fastapi==0.110.0
uvicorn[standard]==0.29.0

# HTTP 客户端
httpx==0.27.0

# 数据验证
pydantic==2.7.0

# 数据库 ORM
sqlalchemy==2.0.44

# 认证和加密
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
cryptography==46.0.3

# 文件上传
python-multipart==0.0.20

# WebSocket 支持
websockets==15.0.1

# 定时任务
apscheduler==3.10.4

# 浏览器自动化（用于保活任务）
selenium==4.27.1
